Time Input Works & Has Awesome Animations
The Time Input validates if the user inputs a valid time and also creates awesome animations when it's hovered or when the user changes the time's value. It also displays the selected hour if the user has left the input and the time is a valid one. But, if the selected time is an invalid one, it shows a message in the input indicating the user to insert another time