Date Input Works
The date input has already been completed. You can choose an appointment day starting from the day you consult the website up to the next 3 months. You can only select weekdays and the wrong and correct effect in the input works fine. The only thing to modify is the position of the text inside the input, I have to change the padding-top of the label for some screen-widths