Autocomplete Problem has been solved
When you want to type your First Name or your Last Name, the browser can give you possible choices for autocomplete your name, based on previous text inputs you have completed. But when the user utilizes this autocomplete option, the style of the input changes. So, in order to maintain the styles applied to the other inputs like the date or time input, I have changed the styles for the autocomplete option with CSS, the code is in the file named '_autocomplete.scss' and some code has been added into the 'formAnimations.js' file