Modernizr File added
modernizr.js File has been added. The purpose of this file is to know whether the browser accepts '.webp' images or not. If the browser doesn't accept '.webp' images, then the page will load a '.jpg' image in order to make this site responsive. And also, it is a test for knowing if a change is made on the Github Page for this project when i make a pull into the Git repository